dev-requirements.txt
decorator==4.0.10
ipython==5.0.0
ipython-genutils==0.1.0
odfpy==1.3.2
pep8==1.7.0
pexpect==4.2.0
pickleshare==0.7.2
pkg-resources==0.0.0
prompt-toolkit==1.0.3
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pytest==2.9.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
